+++ CHANGELOG.md
@@ -9,7 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- A new parameter `FixedPreviewAspectRatio` (boolean) to `settings.cfg`, allowing the in-game camera preview to be forced to a 1:1 aspect ratio
 - Mod version label in settings menu
+
+### Fixed
+
+- Drag the in-game preview by the camera image (on large screens, it can be difficult to target the title bar with the mouse...)
+- More space for the streaming label
 
 ## v2.1.0 - 2026-04-22